refactor: delete 186 unreachable frontend sources and add the clippy ratchet
Wave W4a. Behaviour-preserving: these files cannot run.

Phase 1's reachability scan found 191 of 428 tracked frontend sources unreachable from the
three Vite entries. This commit removes 186 of them, in 3 batches, each verified by a
production `vite build` before the next began:

  icons + date-picker + atoms   103 files
  molecules + ui + search-modal  59 files
  libs + types + hooks + assets  24 files

Safety checks before deleting anything:
- A cross-reference scan confirmed 0 live files import an unreachable one.
- The 191 figure was audited rather than trusted. Five were false positives in the scanner
  and are now correctly treated as live: src/lib/i18n-vite-loaded/** (imported by
  vite.config.mts as a build plugin) and the ambient .d.ts files (pulled in by tsconfig
  include, never by an import). reachability.mjs learned both cases so a future delete wave
  cannot remove build tooling.
- reachability.mjs --count now reports 0, and the production build passes at every step.

Why this wave exists: the plan's original wave list went straight from backend splits to
frontend splits, but splitting a 2000-line NavBar while 186 dead files sit in the same
module graph is wasted effort — and it left the count of dead files invisible.

Effect on the gates: TypeScript errors fall from 408 to 305. The remainder is ~301 vendored
(react-twitter-embed cypress tests, excluded permanently) plus a small tail in reachable
files, so the typecheck gate is now close to its exit condition.

Clippy: the crate emitted 647 warnings, which `-D warnings` cannot gate on. Measured and
split — 364 are rustc unexpected_cfgs and 124 deprecated (dependency/toolchain noise), and
only 62 are actual clippy lints in our own targets. `cargo clippy --fix` applied 236
mechanical fixes (298 -> 62 clippy lints), and scripts/harness/clippy-ratchet.mjs now
enforces the 62 as a baseline that may only shrink. No #[allow] was added anywhere.

rustfmt moves to scripts/harness/check-rustfmt.sh: src/schema.rs is generated with 4-space
indentation that conflicts with tab_spaces = 2, and rustfmt has no stable way to exclude one
path (rustfmt.toml `ignore` needs nightly, #![rustfmt::skip] is unstable, and rustfmt
follows `mod schema;` into it regardless). Files are checked individually so the generated
file can be skipped; vendored libs are excluded per GOLDEN-RULES R7.

ISSUE-004/005 hygiene fix actually lands here: .env, pastebar_settings.yaml, the Vite
timestamp-*.mjs temp file and both generated tailwind-safelist.txt copies are untracked
with working copies preserved. Verified first that the app never reads .env at runtime.

Gates: 10/10 pass (check-all.sh). Refs HARNESS_PLAN.md phase 4 (wave W4a), ISSUE-030.

+++ src-tauri/src/window_ext.rs
@@ -77,7 +77,7 @@ impl<R: Runtime> WindowToolBar for Window<R> {
 
       let window_buttons = vec![close, miniaturize, zoom];
       let space_between =
-        NSView::frame(miniaturize).origin.x - NSView::frame(close).origin.x + 20 as f64;
+        NSView::frame(miniaturize).origin.x - NSView::frame(close).origin.x + 20_f64;
 
       for (i, button) in window_buttons.into_iter().enumerate() {
         let mut rect: NSRect = NSView::frame(button);
